Circuit to STG conv. with/without env. composition
Closes #744
