+++ integration-tests/convert-circuit/test.sh
@@ -1,0 +1,5 @@
+script_file=${test_dir}/vme-tm.circuit.js
+
+[[ -e $script_file ]] || error "Script file ${script_file} is missing"
+
+./workcraft -nogui -dir:${test_dir} -exec:${script_file} >${log_file}
